+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 ### To be Released
 
-* Update Scalingo internal dependencies to the Go Modules version []()
+* Update various dependencies:
+  - github.com/fatih/color
+  - github.com/briandowns/spinner
+  - github.com/gosuri/uilive
+  - github.com/stretchr/testify
+  - github.com/urfave/cli
+  - gopkg.in/AlecAivazis/survey.v1
+  - github.com/cheggaaa/pb
+  - gopkg.in/src-d/go-git.v4
+* Update Scalingo internal dependencies to the Go Modules version [#613](https://github.com/Scalingo/cli/pull/613)
 * Show the addon status on scalingo addons [#604](https://github.com/Scalingo/cli/pull/604)
 * Add informative error in case of container type error when scaling an application
   [602](https://github.com/Scalingo/cli/pull/602)