+++ CHANGELOG.md
@@ -19,7 +19,7 @@
   - golang.org/x/crypto
   - golang.org/x/net
 * Update Scalingo internal dependencies to the Go Modules version [#613](https://github.com/Scalingo/cli/pull/613)
-* Show the addon status on scalingo addons [#604](https://github.com/Scalingo/cli/pull/604)
+* Show the addon status on `scalingo addons` [#604](https://github.com/Scalingo/cli/pull/604)
 * Add informative error in case of container type error when scaling an application
   [602](https://github.com/Scalingo/cli/pull/602)